Route navigation checks user privilege level *before* routing
