+++ ZelFront/src/navigation/vertical/daemon/wallet.js
@@ -3,5 +3,6 @@ export default [
     title: 'Get Wallet Info',
     icon: 'wallet',
     route: 'daemon-wallet-getwalletinfo',
+    privilege: ['user', 'admin', 'fluxteam'],
   },
 ]
