+++ ZelFront/src/router/index.js
@@ -1,5 +1,9 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
+import store from '@/store'
+
+import IDService from '@/services/IDService'
+
 import dashboard from './routes/dashboard'
 import daemon from './routes/daemon/daemon'
 import benchmark from './routes/benchmark/benchmark'
@@ -7,6 +11,8 @@ import flux from './routes/flux/flux'
 import apps from './routes/apps/apps'
 import fluxadmin from './routes/fluxadmin/fluxadmin'
 import xdao from './routes/xdao/xdao'
+
+const qs = require('qs')
 
 Vue.use(VueRouter)
 
@@ -65,6 +71,34 @@ const router = new VueRouter({
       redirect: 'error-404',
     },
   ],
+})
+
+router.beforeEach(async (to, from, next) => {
+  const zelidauth = localStorage.getItem('zelidauth')
+  const auth = qs.parse(zelidauth)
+  store.commit('flux/setPrivilege', 'none')
+  if (auth && auth.zelid && auth.signature) {
+    try {
+      const response = await IDService.checkUserLogged(auth.zelid, auth.signature)
+      const privilege = response.data.data.message
+      store.commit('flux/setPrivilege', privilege)
+      if (privilege === 'none') {
+        localStorage.removeItem('zelidauth')
+      }
+    } catch (error) {
+      console.log(error)
+    }
+  }
+
+  if (to.meta && to.meta.privilege) {
+    if (to.meta.privilege.some(value => value === store.state.flux.privilege)) {
+      next()
+    } else {
+      next('/')
+    }
+  } else {
+    next()
+  }
 })
 
 // ? For splash screen
