+++ ZelFront/src/router/routes/apps/apps.js
@@ -65,6 +65,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin'],
     },
   },
 ]
