+++ ZelFront/src/router/routes/benchmark/benchmark.js
@@ -21,6 +21,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin', 'fluxteam'],
     },
   },
 ]
