+++ ZelFront/src/router/routes/benchmark/benchmarks.js
@@ -37,6 +37,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin', 'fluxteam'],
     },
   },
   {
@@ -57,6 +58,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin'],
     },
   },
 ]
