+++ ZelFront/src/router/routes/daemon/control.js
@@ -57,6 +57,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin'],
     },
   },
   {
@@ -77,6 +78,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin'],
     },
   },
   {
@@ -97,6 +99,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin', 'fluxteam'],
     },
   },
   {
@@ -117,6 +120,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin'],
     },
   },
   {
@@ -137,6 +141,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin', 'fluxteam'],
     },
   },
 ]
