+++ ZelFront/src/router/routes/daemon/daemon.js
@@ -33,6 +33,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin', 'fluxteam'],
     },
   },
 ]
