+++ ZelFront/src/router/routes/daemon/wallet.js
@@ -14,6 +14,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['user', 'admin', 'fluxteam'],
     },
   },
 ]
