+++ ZelFront/src/router/routes/fluxadmin/fluxadmin.js
@@ -14,6 +14,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin', 'fluxteam'],
     },
   },
   {
@@ -31,6 +32,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin', 'fluxteam'],
     },
   },
   {
@@ -48,6 +50,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin', 'fluxteam'],
     },
   },
   {
@@ -65,6 +68,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin', 'fluxteam'],
     },
   },
   {
@@ -82,6 +86,7 @@ export default [
           active: true,
         },
       ],
+      privilege: ['admin', 'fluxteam'],
     },
   },
 ]
